by nielsvandermolen: Refactor activity stream modules to be more separate
