+++ public_html/profiles/social/modules/custom/activity_basics/activity_basics.info.yml
@@ -1,0 +1,27 @@
+name: Activity Basics
+type: module
+description: Basic implementation of activites in Open Social
+core: 8.x
+package: Custom
+dependencies:
+  - activity_creator
+  - activity_logger
+  - activity_viewer
+  - block_content
+  - comment
+  - dynamic_entity_reference
+  - field
+  - file
+  - group
+  - menu_link_content
+  - message
+  - node
+  - options
+  - profile
+  - social_event
+  - social_group
+  - social_post
+  - taxonomy
+  - text
+  - user
+version: 1.0
