+++ public_html/profiles/social/modules/custom/activity_basics/activity_basics.module
@@ -1,0 +1,20 @@
+<?php
+
+/**
+ * @file
+ * Contains activity_basics.module..
+ */
+
+use Drupal\Core\Entity\EntityInterface;
+
+/**
+ * Implements hook_ENTITY_TYPE_insert().
+ */
+function activity_logger_entity_insert(EntityInterface $entity) {
+  /* @var $activity_logger_factory Drupal\activity_creator\Plugin\ActivityActionManager */
+  $activity_logger_factory = \Drupal::service('plugin.manager.activity_action.processor');
+  // Trigger the create action for enttites.
+  /* @var $create_action Drupal\activity_creator\Plugin\ActivityAction\CreateActivityAction */
+  $create_action = $activity_logger_factory->createInstance('create_entitiy_action');
+  $create_action->create($entity);
+}
