+++ public_html/profiles/social/modules/custom/activity_basics/config/install/block.block.views_block__activity_stream_group_block_stream_group.yml
