+++ public_html/profiles/social/modules/custom/activity_basics/config/install/core.entity_view_display.activity.activity.notification_archive.yml
