+++ public_html/profiles/social/modules/custom/activity_basics/config/install/views.view.activity_stream_group.yml
