+++ public_html/profiles/social/modules/custom/activity_basics/src/Plugin/ActivityAction/CreateActivityAction.php
@@ -2,10 +2,10 @@
 
 /**
  * @file
- * Contains \Drupal\activity_creator\Plugin\ActivityAction\CreateActivityAction.
+ * Contains \Drupal\activity_basics\Plugin\ActivityAction\CreateActivityAction.
  */
 
-namespace Drupal\activity_creator\Plugin\ActivityAction;
+namespace Drupal\activity_basics\Plugin\ActivityAction;
 
 use Drupal\activity_creator\Plugin\ActivityActionBase;
