+++ public_html/profiles/social/modules/custom/activity_basics/src/Plugin/ActivityContext/ContentInMyGroupContext.php
@@ -2,30 +2,31 @@
 
 /**
  * @file
- * Contains \Drupal\activity_creator\Plugin\ActivityContext\GroupActivityContext.
+ * Contains \Drupal\activity_basics\Plugin\ActivityContext\ContentInMyGroupActivityContext.
  */
 
-namespace Drupal\activity_creator\Plugin\ActivityContext;
+namespace Drupal\activity_basics\Plugin\ActivityContext;
 
 use Drupal\activity_creator\Plugin\ActivityContextBase;
+use Drupal\group\Entity\Group;
 use Drupal\group\Entity\GroupContent;
+use Drupal\group\GroupMembership;
 use Drupal\social_group\SocialGroupHelperService;
 
 /**
- * Provides a 'GroupActivityContext' activity context.
+ * Provides a 'ContentInMyGroupActivityContext' acitivy context.
  *
  * @ActivityContext(
- *  id = "group_activity_context",
- *  label = @Translation("Group activity context"),
+ *  id = "content_in_my_group_activity_context",
+ *  label = @Translation("Content in my group activity context"),
  * )
  */
-class GroupActivityContext extends ActivityContextBase {
+class ContentInMyGroupActivityContext extends ActivityContextBase {
 
   /**
    * {@inheritdoc}
    */
   public function getRecipients(array $data, $last_uid, $limit) {
-
     $recipients = [];
 
     // We only know the context if there is a related object.
@@ -38,11 +39,20 @@ class GroupActivityContext extends ActivityContextBase {
           'target_type' => 'group',
           'target_id' => $gid,
         ];
+        $group = Group::load($gid);
+        $memberships = GroupMembership::loadByGroup($group);
+        foreach ($memberships as $membership) {
+          $recipients[] = [
+            'target_type' => 'user',
+            'target_id' => $membership->getUser()->id(),
+          ];
+        }
       }
     }
 
     return $recipients;
   }
+
 
   public function isValidEntity($entity) {
     // Check if it's placed in a group (regardless off content type).
