+++ public_html/profiles/social/modules/custom/activity_basics/src/Plugin/ActivityDestination/NotificationsActivityDestination.php
@@ -2,10 +2,10 @@
 
 /**
  * @file
- * Contains \Drupal\activity_creator\Plugin\ActivityDestination\NotificationsActivityDestination.
+ * Contains \Drupal\activity_basics\Plugin\ActivityDestination\NotificationsActivityDestination.
  */
 
-namespace Drupal\activity_creator\Plugin\ActivityDestination;
+namespace Drupal\activity_basics\Plugin\ActivityDestination;
 use Drupal\activity_creator\Plugin\ActivityDestinationBase;
 
 /**
