+++ public_html/profiles/social/modules/custom/activity_basics/src/Plugin/ActivityDestination/StreamProfileActivityDestination.php
@@ -2,10 +2,10 @@
 
 /**
  * @file
- * Contains \Drupal\activity_creator\Plugin\ActivityDestination\StreamProfileActivityDestination.
+ * Contains \Drupal\activity_basics\Plugin\ActivityDestination\StreamProfileActivityDestination.
  */
 
-namespace Drupal\activity_creator\Plugin\ActivityDestination;
+namespace Drupal\activity_basics\Plugin\ActivityDestination;
 use Drupal\activity_creator\Plugin\ActivityDestinationBase;
 
 /**
