+++ public_html/profiles/social/modules/custom/activity_creator/config/install/field.storage.activity.field_activity_entity.yml
@@ -3,19 +3,7 @@ status: true
 dependencies:
   module:
     - activity_creator
-    - block_content
-    - comment
     - dynamic_entity_reference
-    - file
-    - group
-    - menu_link_content
-    - message
-    - node
-    - profile
-    - social_event
-    - social_post
-    - taxonomy
-    - user
 id: activity.field_activity_entity
 field_name: field_activity_entity
 entity_type: activity
