+++ public_html/profiles/social/modules/custom/activity_viewer/activity_viewer.info.yml
@@ -1,0 +1,22 @@
+name: Activity Viewer
+type: module
+description: Used to display activities with support for views
+core: 8.x
+package: Social
+dependencies:
+  - activity_creator
+  - activity_logger
+  - block_content
+  - comment
+  - dynamic_entity_reference
+  - file
+  - group
+  - menu_link_content
+  - message
+  - node
+  - options
+  - profile
+  - search
+  - social_post
+  - taxonomy
+  - user
