+++ public_html/profiles/social/modules/custom/activity_viewer/src/Plugin/views/filter/ActivityPostVisibilityAccess.php
@@ -4,7 +4,7 @@
  * Contains \Drupal\activity_creator\Plugin\views\filter\ActivityPostVisibilityAccess.
  */
 
-namespace Drupal\activity_creator\Plugin\views\filter;
+namespace Drupal\activity_viewer\Plugin\views\filter;
 
 use Drupal\views\Plugin\views\filter\FilterPluginBase;
 use Drupal\views\Views;
