+++ public_html/profiles/social/modules/custom/activity_viewer/src/Plugin/views/filter/ActivityVisibilityAccess.php
@@ -5,7 +5,7 @@
  * Contains \Drupal\activity_creator\Plugin\views\filter\ActivityVisibilityAccess.
  */
 
-namespace Drupal\activity_creator\Plugin\views\filter;
+namespace Drupal\activity_viewer\Plugin\views\filter;
 
 use Drupal\views\Plugin\views\filter\FilterPluginBase;
