+++ public_html/profiles/social/modules/social_features/social_activity/social_activity.info.yml
@@ -3,8 +3,7 @@ description: 'Activity feature for Open Social'
 type: module
 core: 8.x
 dependencies:
-  - activity_creator
-  - activity_logger
+  - activity_basics
   - message
   - token
 package: Social
